+++ controllers/engine/DriverPersona/reserveName.js
@@ -1,13 +1,14 @@
-const Users = require('../../../models/user');
+const Personas = require('../../../models/persona');
 
-const nameFinder = async (req, res, next) => {
-  try {
-    const persona = await Users.findOne({ 'persona.name': req.query.name });
-    const available = !persona;
-    res.type('application/xml').render('handlers/string.ejs', { flag: available });
-  } catch (err) {
-    next(err);
-  }
+const nameFinder = async (req, res) => {
+  const { query: { name }} = req;
+  
+  const persona = await Personas.findOne({ name });
+  const available = !persona;
+  
+  return res
+    .type('application/xml')
+    .render('handlers/string.ejs', { flag: available });
 };
 
 module.exports = nameFinder;