+++ controllers/engine/User/permSess.js
@@ -1,9 +1,9 @@
-const Users = require('../../../models/user');
+const Personas = require('../../../models/persona');
 
 const permSess = async (req, res, next) => {
   try {
     const { decoded: { id }, token } = req;
-    const { persona } = await Users.findOne({ _id: id });
+    const persona = await Personas.find({ userId: id });
 
     const data = { id, token, persona };
     return res.render('user/session', { data });