+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 23-05-2021
+
+* Separated personas from user model to separate model.
 # 13-09-2020
 
 * Created changelog and migrated all previous code under this date